[Behat] Extracted countries creating logic from ChannelContext to GeographicalContext
